feat(hooks): refactor SideBarMenu, add custom hook to handle closing with Esc & click
