+++ src/components/SideBar/SideBar.tsx
@@ -1,20 +1,20 @@
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 /* eslint-disable jsx-a11y/click-events-have-key-events */
-import React, { useEffect, useState } from 'react';
-import SideBarUser from 'components/SideBarUser/SideBarUser';
-import { v4 as uuidv4 } from 'uuid';
-import { lid } from 'assets/images'; // Initial image
 import {
-  allTaskIcon,
-  requestIcon,
-  myTaskIcon,
-  archiveIcon,
-  analyticsIcon,
   AllTeamIcon,
+  allTaskIcon,
+  analyticsIcon,
+  archiveIcon,
+  myTaskIcon,
+  requestIcon,
 } from 'assets/icons';
+import { lid } from 'assets/images'; // Initial image
+import SideBarUser from 'components/SideBarUser/SideBarUser';
 import SideBarUserMenu from 'components/SideBarUserMenu/SideBarUserMenu';
+import React, { useState } from 'react';
 import { useSelector } from 'services/hooks';
 import { TTask } from 'types/types';
+import { v4 as uuidv4 } from 'uuid';
 import styles from './SideBar.module.scss';
 
 const SideBar: React.FC = () => {
@@ -28,28 +28,12 @@ const SideBar: React.FC = () => {
     setisSidebarMenuOpen(!isSidebarMenuOpen);
   };
 
-  const handleCloseMenu = () => {
-    setisSidebarMenuOpen(false);
-  };
-
-  const handleEscapeClick = (evt: KeyboardEvent) => {
-    if (evt.key === 'Escape') {
-      handleCloseMenu();
-    }
-  };
-  useEffect(() => {
-    document.addEventListener('keydown', handleEscapeClick);
-    return () => {
-      document.removeEventListener('keydown', handleEscapeClick);
-    };
-  });
-
   return (
     <section className={styles.SideBar}>
       <div onClick={handleToggleMenu} className={styles.userWrapper}>
         <img className={styles.userImg} src={lid} alt="Изображение пользователя" />
         <h2 className={styles.userName}>{currentUser.full_name}</h2>
-        <SideBarUserMenu isOpen={isSidebarMenuOpen} />
+        {isSidebarMenuOpen && <SideBarUserMenu setIsOpen={setisSidebarMenuOpen} />}
       </div>
       <ul className={styles.linkWrapper}>
         <li className={`${styles.navLi} ${styles.navLiActive}`}>
